Reorganize TCPServer startup/shutdown
This prevents a race condition on bundle refreshing. The problem
was that TCPServer's activate() and deactivate() methods were written
to return asynchronously, which meant that in the case of deactivate(),
the server socket may not have been closed by the time the new instance
of TCPServer was started. The new instance would then fail to bind a
socket to the same address, because the old one was still open. Now,
the methods work synchronously and the problem cannot occur.